Refactored how arrow pointing works in tutorial
There is no longer one arrow that moves around to different locations (this was very temperamental). Using specific coordinates won't work if the UI is scaling, and while I attemped to move around an arrow based on hidden gameobjects at relevant locations this was also too temperamental. Ended up duplicating the arrow image and moving them to their correct locations, after which I refactored the code so that each game object is deactivated and activated as it moves on to each stage.